set dots color
when an answer is being clicked, all dots (which represent the participants answers) where participants have given the same answer should light up in the same color. Right now it still counts answers such as No for all instances that they were given (e. g. when a no answer was given for any answer from the questionnaire). This bug should be solved.