add infobox for segment on hover
When the user hovers over a segment of the outer ring, an infobox to the side of the diagram changes according to the segment that the cursor moves over. The position and design of the box has to be improved. Also when the mouse doesn't hover over a sgment the text in the info box may change to a default text or disappear entirely. To make room for the info box, the diagram was reduced in size and moved more to the left side.