rest inner ring colors
When a user clicks on a dot, the answers given by the participant that correspond to the segments of the inner ring change color, when the user now clicks on an answer, the colors reset and the inner ring segments retain their original color. The information about the color range 8or the function itself) maybe should be moved to startDiagramManager. Outsorcing the visual changes should be considered after making sure that all remaining important tasks for the start diagram are done