select multiple answers
The user can now select multiple answers to change the color of the dots on the display. The active segments get a distinct color to show that they are active and the elements themselves are being saved in an array. If the user clicks on an active segment it will get back its original color and it is removed from the array of active elements. Some additional code comments were made and previous ones were improved or changed to better understand them. Right now the functionaltiy of selecting multiple answers only works with "OR" logic instead of "AND" logic. This will be implemented next.