added zoom functionality
the inner ring now has zoom functionality. There are some problems that still remain. The segments disappear when the user zooms out too far and the segments all move at once. It should be tested if segments could be zoomed separately or if the svg as a whole should be zoomed