improved zoom
The zoom now happens at the same place where the cursor is whlile the user uses the mouse wheel or touch gestures to zoom. If the inner ring is offset after the zoom, the user can move the inner ring to its original position with drag and drop. The info box at the Home page is now smaller because there is no additional information being displayed for the participant dots. The size of the other answers was improved to fill more space and improve readability