move/rewrite view code
the view code inside the controller was moved to the startDiagramModel, which handles the view functionality and the general visuals for the start diagram. This allows for a better use of the MVC pattern. Some code was rewritten to work with this new division. Furthermore unnecessary code was removed and other segments were shortened  to improve readability.